+++ docs/changelog.rst
@@ -45,6 +45,8 @@ Also:
 
 * Added compatibility with Python 3.7.
 
+* Added an interactive client: `python -m websockets <uri>`
+
 5.0
 ...